Review. Review fixes

+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -3,12 +3,12 @@ package ru.practicum.shareit.user;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.practicum.shareit.user.dto.UserCreateDto;
+import ru.practicum.shareit.user.dto.UserDtoOut;
 import ru.practicum.shareit.user.dto.UserUpdateDto;
 import ru.practicum.shareit.user.exception.UserEmailAlreadyExist;
 import ru.practicum.shareit.user.exception.UserNotFoundException;
 
 import java.util.Collection;
-import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
@@ -17,24 +17,26 @@ public class UserService {
     private final UserDao userDao;
     private final UserMapper userMapper;
 
-    public User addUser(UserCreateDto dto) {
+    public UserDtoOut addUser(UserCreateDto dto) {
         checkUniqueEmail(dto.getEmail());
         User user = userMapper.map(dto);
-        return userDao.save(user);
+        userDao.save(user);
+        return userMapper.map(user);
     }
 
     public void deleteUser(long userId) {
-        userDao.findByUserId(userId)
-                .orElseThrow(() -> new UserNotFoundException("User is not found"));
+        if (!userDao.contains(userId)) {
+            throw new UserNotFoundException("User is not found");
+        }
         userDao.deleteByUserId(userId);
     }
 
-    public User updateUser(UserUpdateDto dto, long userId) {
+    public UserDtoOut updateUser(UserUpdateDto dto, long userId) {
         final User user = userDao.findByUserId(userId)
                 .orElseThrow(() -> new UserNotFoundException("User is not found"));
         if ((dto.getEmail() == null || dto.getEmail().equals(user.getEmail()))
                 && (dto.getName() == null || dto.getName().equals(user.getName()))) {
-            return user;
+            return userMapper.map(user);
         }
         if (dto.getEmail() != null && !dto.getEmail().equals(user.getEmail())) {
             checkUniqueEmail(dto.getEmail());
@@ -43,21 +45,23 @@ public class UserService {
         if (dto.getName() != null && !dto.getName().equals(user.getName())) {
             user.setName(dto.getName());
         }
-        return userDao.save(user);
+        userDao.save(user);
+        return userMapper.map(user);
     }
 
-    public User getUser(long userId) {
-        return userDao.findByUserId(userId)
+    public UserDtoOut getUser(long userId) {
+        User user = userDao.findByUserId(userId)
                 .orElseThrow(() -> new UserNotFoundException("User is not found"));
+        return userMapper.map(user);
     }
 
-    public Collection<User> getAllUsers() {
-        return userDao.findAll();
+    public Collection<UserDtoOut> getAllUsers() {
+        Collection<User> users = userDao.findAll();
+        return userMapper.map(users);
     }
 
     private void checkUniqueEmail(String email) {
-        Set<String> emails = userDao.findAllEmail();
-        if (emails.contains(email)) {
+        if (userDao.existByEmail(email)) {
             throw new UserEmailAlreadyExist(String.format("User email %s already exist", email));
         }
     }
